docs: describe your change...
I have been in trouble following up the tutorial becasue is not saving my previous updates and there's no any branch state mentioned that could it be used lesson by lesson. So I'm adding the note to start forking the project so You can save the progress of the updates You do over the tutorial.